refactor: 패키지 분리
- Board → chess 패키지
- Pawn → pieces 패키지
