+++ src/test/java/woowa/camp/chess/BoardTest.java
@@ -1,10 +1,11 @@
-package woowa.camp;
+package woowa.camp.chess;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import woowa.camp.pieces.Pawn;
 
 public class BoardTest {
